front/requirements.txt
cassandra-driver==3.25.0
Flask==2.2.2
python-dotenv==0.21.0
